refactor: next.config.mjs -> next.config.ts
