+++ tsconfig.json
@@ -33,7 +33,7 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    "next.config.mjs"
+    "next.config.js"
   ],
   "exclude": ["node_modules"]
 }
